+++ src/components/Layout.tsx
@@ -7,7 +7,7 @@ export default function Layout({ children } : { children: ReactNode}) {
   return (
     <main className="p-4">
       <div className="grid grid-rows-1 lg:grid-cols-12 md:grid-cols-1 gap-4">
-        <div className="col-span-3">
+        <div className="col-span-3 border-r-[1px] border-solid">
           <SideBar />
         </div>
         <div className="col-span-9">
